+++ src/inquirer/messageInteractionComponent.ts
